requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
